refactor: Money를 객체로 포장
