+++ src/main/java/lotto/controller/LottoController.java
@@ -2,12 +2,13 @@ package lotto.controller;
 
 import lotto.model.LottoGame;
 import lotto.model.Lottos;
+import lotto.model.Money;
 import lotto.view.InputView;
 import lotto.view.ResultView;
 
 public class LottoController {
     public static void runGame() {
-        Integer money = InputView.insertMoney();
+        Money money = new Money(InputView.insertMoney());
         Lottos lottos = LottoGame.buyLottos(money);
         ResultView.printGeneratedLottos(lottos.getLottos());
