+++ src/test/java/lotto/model/LottoGameTest.java
@@ -16,29 +16,12 @@ class LottoGameTest {
     @DisplayName("수익률 계산 확인")
     void calculateYieldTest() {
         LottoGame lottoGame = new LottoGame(Arrays.asList(1, 2, 3, 4, 5, 6), 7);
-        int money = 14000;
+        Money money = new Money(14000);
         Lottos lottos = new Lottos(Collections.singletonList(new Lotto(Arrays.asList(1, 2, 3, 11, 12, 13))));
         lottoGame.generateLottoResult(lottos);
         float yield = lottoGame.calculateYield(money);
 
         assertThat(yield).isCloseTo(0.35f, Percentage.withPercentage(99));
-    }
-
-    @Test
-    @DisplayName("로또 구입 금액 입력시 몇 장의 로또 생성하는지 확인")
-    void generateLottosTest() {
-        int numberOfLottos = LottoGame.getLottoSize(14000);
-
-        assertThat(numberOfLottos).isEqualTo(14);
-    }
-
-    @Test
-    @DisplayName("로또 구매 금액이 로또 가격으로 나눠떨어지지 않으면 예외를 던진다")
-    void validateUnitPrice() {
-        assertThatThrownBy(() -> {
-            LottoGame.getLottoSize(14100);
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("거스름돈을 지급하지 않습니다. 금액이 남지 않게 지불해주세요.");
     }
 
     @Test
